+++ ibtk/src/utilities/ParallelEdgeMap.cpp
@@ -14,8 +14,8 @@
 //      notice, this list of conditions and the following disclaimer in the
 //      documentation and/or other materials provided with the distribution.
 //
-//    * Neither the name of New York University nor the names of its
-//      contributors may be used to endorse or promote products derived from
+//    * Neither the name of The University of North Carolina nor the names of
+//      its contributors may be used to endorse or promote products derived from
 //      this software without specific prior written permission.
 //
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
@@ -35,9 +35,10 @@
 #include <algorithm>
 #include <map>
 #include <ostream>
+#include <utility>
 #include <vector>
 
-#include "ParallelEdgeMap.h"
+#include "ibtk/ParallelEdgeMap.h"
 #include "ibtk/namespaces.h" // IWYU pragma: keep
 #include "tbox/SAMRAI_MPI.h"
 #include "tbox/Utilities.h"
@@ -94,8 +95,8 @@ void ParallelEdgeMap::communicateData()
     const int rank = SAMRAI_MPI::getRank();
 
     std::vector<int> num_additions_and_removals(2 * size, 0);
-    num_additions_and_removals[2 * rank] = d_pending_additions.size();
-    num_additions_and_removals[2 * rank + 1] = d_pending_removals.size();
+    num_additions_and_removals[2 * rank] = static_cast<int>(d_pending_additions.size());
+    num_additions_and_removals[2 * rank + 1] = static_cast<int>(d_pending_removals.size());
     SAMRAI_MPI::sumReduction(&num_additions_and_removals[0], 2 * size);
 
     int num_transactions = 0, offset = 0;
@@ -113,8 +114,7 @@ void ParallelEdgeMap::communicateData()
 
     static const int SIZE = 3;
     std::vector<int> transactions(SIZE * num_transactions, 0);
-    for (std::multimap<int, std::pair<int, int> >::const_iterator
-             cit = d_pending_additions.begin();
+    for (std::multimap<int, std::pair<int, int> >::const_iterator cit = d_pending_additions.begin();
          cit != d_pending_additions.end();
          ++cit, ++offset)
     {
@@ -122,8 +122,7 @@ void ParallelEdgeMap::communicateData()
         transactions[SIZE * offset + 1] = cit->second.first;
         transactions[SIZE * offset + 2] = cit->second.second;
     }
-    for (std::multimap<int, std::pair<int, int> >::const_iterator
-             cit = d_pending_removals.begin();
+    for (std::multimap<int, std::pair<int, int> >::const_iterator cit = d_pending_removals.begin();
          cit != d_pending_removals.end();
          ++cit, ++offset)
     {
@@ -156,25 +155,20 @@ void ParallelEdgeMap::communicateData()
 
     typedef std::multimap<int, std::pair<int, int> >::iterator multimap_iterator;
     typedef std::multimap<int, std::pair<int, int> >::const_iterator multimap_const_iterator;
-    for (multimap_const_iterator cit = d_pending_additions.begin();
-         cit != d_pending_additions.end();
-         ++cit)
+    for (multimap_const_iterator cit = d_pending_additions.begin(); cit != d_pending_additions.end(); ++cit)
     {
         d_edge_map.insert(std::make_pair(cit->first, cit->second));
     }
 
     typedef std::multimap<int, std::pair<int, int> >::const_iterator multimap_const_iterator;
-    for (multimap_const_iterator cit = d_pending_removals.begin();
-         cit != d_pending_removals.end();
-         ++cit)
+    for (multimap_const_iterator cit = d_pending_removals.begin(); cit != d_pending_removals.end(); ++cit)
     {
         int mastr_idx = cit->first;
         const std::pair<int, int>& link = cit->second;
 
         bool found_link = false;
 
-        std::pair<multimap_iterator, multimap_iterator> range =
-            d_edge_map.equal_range(mastr_idx);
+        std::pair<multimap_iterator, multimap_iterator> range = d_edge_map.equal_range(mastr_idx);
         for (multimap_iterator it = range.first; it != range.second && !found_link; ++it)
         {
             if (it->second == link)
@@ -197,8 +191,7 @@ void ParallelEdgeMap::communicateData()
                 mastr_idx = idx1;
             }
 
-            std::pair<multimap_iterator, multimap_iterator> range =
-                d_edge_map.equal_range(mastr_idx);
+            std::pair<multimap_iterator, multimap_iterator> range = d_edge_map.equal_range(mastr_idx);
             for (multimap_iterator it = range.first; it != range.second && !found_link; ++it)
             {
                 if (it->second == link)
